requirements.txt
robotframework==4.0
robotframework-seleniumlibrary==6.1.0
robotframework-jsonlibrary